+++ src/pages/Profile/Invitations/Styles.ts
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const Container = styled.div``;
+
+export const InvitationContainer = styled.div`
+  display: grid;
+  gap: 12px;
+  grid-template-columns: 2fr 2fr 1fr 1fr min-content;
+  padding: 12px 0;
+`;
