+++ src/pages/Profile/index.tsx
@@ -1,26 +1,18 @@
-import {useEffect, useMemo} from 'react';
+import {useEffect} from 'react';
 import {useDispatch, useSelector} from 'react-redux';
-import orderBy from 'lodash/orderBy';
 
-import Button from 'components/Button';
-import EmptyText from 'components/EmptyText';
-import SectionHeading from 'components/SectionHeading';
 import {getInvitationLimit} from 'dispatchers/invitationLimits';
 import {getInvitations as _getInvitations} from 'dispatchers/invitations';
 import {useSelfAvatar, useToggle} from 'hooks';
 import EditProfileModal from 'modals/EditProfileModal';
-import InvitationModal from 'modals/InvitationModal';
-import {getInvitationLimits, getInvitations, getSelf} from 'selectors/state';
+import {getSelf} from 'selectors/state';
 import {AppDispatch, SFC} from 'types';
-import Invitation from './Invitation';
+import Invitations from './Invitations';
 import * as S from './Styles';
 
 const Profile: SFC = ({className}) => {
   const [editProfileModalIsOpen, toggleEditProfileModal] = useToggle(false);
-  const [invitationModalIsOpen, toggleInvitationModal] = useToggle(false);
   const dispatch = useDispatch<AppDispatch>();
-  const invitationLimits = useSelector(getInvitationLimits);
-  const invitations = useSelector(getInvitations);
   const self = useSelector(getSelf);
   const selfAvatar = useSelfAvatar();
 
@@ -31,17 +23,6 @@ const Profile: SFC = ({className}) => {
     })();
   }, [dispatch, self.id]);
 
-  const invitationLimitAmount = useMemo(() => {
-    const invitationLimit = Object.values(invitationLimits).find(
-      (_invitationLimit) => _invitationLimit.owner === self.id,
-    );
-    return invitationLimit?.amount || 0;
-  }, [invitationLimits, self.id]);
-
-  const invitationList = useMemo(() => {
-    return Object.values(invitations);
-  }, [invitations]);
-
   const renderAvatar = () => {
     return (
       <S.ImgWrapper>
@@ -51,31 +32,7 @@ const Profile: SFC = ({className}) => {
   };
 
   const renderContent = () => {
-    if (!!invitationList.length) return renderInvitations();
-    return <EmptyText>No invitations to display.</EmptyText>;
-  };
-
-  const renderCreateInvitationButton = () => {
-    if (invitationList.length >= invitationLimitAmount) return null;
-    return <Button onClick={toggleInvitationModal} text="Create" />;
-  };
-
-  const renderInvitations = () => {
-    const orderedInvitations = orderBy(invitationList, ['created_date'], ['desc']);
-    const _invitations = orderedInvitations.map((invitation) => (
-      <Invitation invitation={invitation} key={invitation.id} />
-    ));
-    return <S.Invitations>{_invitations}</S.Invitations>;
-  };
-
-  const renderSectionHeading = () => {
-    return (
-      <SectionHeading
-        heading="Invitations"
-        rightContent={renderCreateInvitationButton()}
-        subHeading={`${invitationList.length}/${invitationLimitAmount}`}
-      />
-    );
+    return <Invitations />;
   };
 
   return (
@@ -85,13 +42,9 @@ const Profile: SFC = ({className}) => {
           {renderAvatar()}
           <S.Button onClick={toggleEditProfileModal} text="Edit Profile" />
         </S.Left>
-        <S.Right>
-          {renderSectionHeading()}
-          {renderContent()}
-        </S.Right>
+        <S.Right>{renderContent()}</S.Right>
       </S.Container>
       {editProfileModalIsOpen ? <EditProfileModal close={toggleEditProfileModal} /> : null}
-      {invitationModalIsOpen ? <InvitationModal close={toggleInvitationModal} /> : null}
     </>
   );
 };
